requirements.txt
pyTelegramBotAPI==4.18.0
openai==0.27.8
python-dotenv==1.0.0
yandex-cloud-ml-sdk>=0.19.0
